# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Hush Hush
  summary: A standalone secrets object store.
  description: |
    A writer only ever needs the service's public key to add or rotate a
    value; a consumer only ever needs an object's slug to fetch it. Every
    object is sealed (age) to one or more recipients chosen at write time -
    the service stores and serves ciphertext only, and never computes or
    returns plaintext.

    v1's read path is deliberately flat: fetching an object's ciphertext
    needs no authorization, only its slug. The confidentiality boundary is
    entirely "who holds a matching private key" - every fetch is recorded
    to the audit log instead, which is what answers "did anyone access
    this" during a rotation or incident. Writes (create, update, delete)
    are gated by a bearer token.

    A separate web UI adds a second, human-facing way in: a single admin
    account authenticates with a WebAuthn passkey and gets a browser
    session, used only by the `/auth/*`, `/credentials/*`, and `/tokens/*`
    endpoints below - it never substitutes for the bearer token the
    `objects`/`audit-log` endpoints above require. WebAuthn needs a fixed
    relying-party ID and origin, so a deployment that enables the web UI
    sets a new `PUBLIC_URL` environment variable (see the server's own
    README) - unset, every `/auth/*` endpoint responds with a
    configuration error rather than guessing a host from request headers.
  version: 0.1.0
  contact:
...
servers:
...
security: []
...
paths:
...
  /objects/{slug}:
    get:
      operationId: getObject
      summary: Fetch an object's sealed ciphertext
      description: |
        Returns the stored ciphertext exactly as sealed, byte for byte -
        the service never decrypts it. Requires a write bearer token, an
        admin session, or a consumer read token whose bound consumer
        appears in this object's `used_by` list - the confidentiality
        boundary is still "who holds a matching private key," this adds
        an access-control layer on top of it
        (openspec/changes/consumer-read-tokens/proposal.md). Every fetch
        is recorded to the audit log regardless of which credential
        authenticated it.

        A consumer token presented for an object outside its scope gets
        the same 404 an unknown slug would, not 403 - so it can't be used
        to enumerate which other slugs exist.
      tags: [objects]
      security:
        - bearerAuth: []
        - cookieAuth: []
        - consumerBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/caller"
      responses:
        "200":
          description: The sealed object.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateObject
      summary: Rotate an object's value
      description: |
        Replaces the stored ciphertext for an existing object. The new
        value is checked the same way a created one is: not base64 of a
        well-formed age file, or no recipient in its header, is a 422
        and the stored value is left alone. The
        object's slug and description metadata are always preserved
        unchanged. used_by is preserved too, unless the request body
        includes it - in which case it fully replaces the object's
        recorded consumers, the same way creating an object sets it. A
        session-authenticated call needs its CSRF token too; a
        bearer-token-authenticated one doesn't.
      tags: [objects]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/caller"
        - $ref: "#/components/parameters/csrfTokenOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateObjectRequest"
            example:
              value: YWdlLWVuY3J5cHRpb24ub3JnL3Yx
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "200":
          description: The object's value was replaced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectMetadata"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteObject
      summary: Remove an object
      description: |
        Permanently removes an object. A subsequent fetch by this slug
        returns 404. A session-authenticated call needs its CSRF token
        too; a bearer-token-authenticated one doesn't.
      tags: [objects]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/caller"
        - $ref: "#/components/parameters/csrfTokenOptional"
      responses:
        "204":
          description: The object was removed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A write token, checked on every create, update, and delete call.
        Each token carries its own description and expiry, issued and
        revoked through the server's own `token` CLI command (direct
        store access, not an HTTP endpoint - issuing a *write* token over
        HTTP would need its own write token to call it with). Any number
        of tokens may be valid at once; revoking or expiring one never
        affects the others.

        Also accepted, as an alternative to `cookieAuth`, on every
        `/consumer-tokens` operation: a write token already reads or
        writes any object, so authorizing it to mint or manage a
        consumer token - itself scoped to reading a single consumer's
        objects - is a narrowing delegation, not the bootstrap problem
        minting a *write* token over HTTP would be
        (alrayyes/hush-hush#467).

    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        A browser session, issued by a successful `/auth/login/finish` or
        `/auth/register/finish` call. Used by the `/auth/*` (post-login),
        `/credentials/*`, and `/tokens/*` endpoints, and as an alternative
        to `bearerAuth` on `/objects/*` - an admin session reads or
        writes any object, the same access a write bearer token has.

    consumerBearerAuth:
      type: http
      scheme: bearer
      description: |
        A consumer read token, scoped to exactly one consumer name at
        creation. Authorizes `getObject` only for an object whose
        `used_by` list includes that token's bound consumer - unlike
        `bearerAuth`, which reads any object. Issued and managed through
        `/consumer-tokens`, never authorizes any endpoint besides
...
  parameters:
    slug:
      name: slug
      in: path
      required: true
      description: |
        The object's slug - the caller-facing identifier a writer chose at
        create time, the last path segment, and so has to survive being in
        a URL. The object's internal id is never part of any path
        (specs/secret-objects/spec.md's "Internal id decoupled from
        user-facing slug" requirement).
      schema:
        $ref: "#/components/schemas/ObjectSlug"
    caller:
      name: X-Caller
      in: header
      required: false
      description: |
        The caller's own identity - a repo or host name, whatever the
        caller wants attributed to it in the audit log. Not authenticated
        or verified: a courtesy label a caller presents about itself, not
        an identity check the service performs. Absent means the
        resulting audit log entry's caller field is empty.
      schema:
        type: string
...
    csrfTokenOptional:
      name: X-CSRF-Token
      in: header
      required: false
      description: |
        The current session's own CSRF token - required when the request
        is authenticated by a session, not present or checked when it's
        authenticated by a bearer token instead (which has no session, and
        so no CSRF token to send). See `csrfToken` for the always-required
        form used where session auth is the only option.
      schema:
        type: string
  responses:
...
    PayloadTooLarge:
      description: |
        The request body is larger than the server reads, or a sealed
        `value` is over its 1 MiB limit. The server stops reading at the
        limit.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: value is larger than the 1048576 byte limit
    UnprocessableEntity:
      description: |
        The request parsed, but a value in it is outside what the API
        allows. The message names the limit or the problem.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: ttl_seconds must be at most 31536000 (365 days)
    Unauthorized:
      description: The request is missing a valid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: missing or invalid bearer token
    NotFound:
      description: No object is stored under that slug.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: unknown object
...
  schemas:
    ObjectSlug:
      type: string
      maxLength: 128
      pattern: "^[a-z0-9][a-z0-9_-]*$"
      examples: [mattermost_deploy_webhook]
      description: |
        A caller-chosen, unique identifier for an object - what every
        documented request path (URL, CLI, API) addresses it by. The
        object's internal id is a separate, opaque value never exposed to
        or accepted from a caller (specs/secret-objects/spec.md's
        "Internal id decoupled from user-facing slug" requirement).
...
    UpdateObjectRequest:
      type: object
      required: [value]
      properties:
        value:
          type: string
          format: byte
          maxLength: 1398104
          description: |
            The new sealed (encrypted) value, base64-encoded. It has to be
            a well-formed age file naming at least one recipient; the
            service reads only its header, never the payload. At most 1 MiB
            once decoded.
        used_by:
          allOf:
            - $ref: "#/components/schemas/UsedByList"
          description: |
            Replaces the object's recorded used_by list, the same way
            CreateObjectRequest's used_by sets it initially - omit this
            field entirely to leave the existing list untouched. An
            empty array clears it.
        tags:
          allOf:
            - $ref: "#/components/schemas/Tags"
          description: |
            Replaces the object's tags - omit this field entirely to leave
            them untouched. An empty array clears them.
        keep_readable_copy:
          $ref: "#/components/schemas/KeepReadableCopy"
    KeepReadableCopy:
      type: boolean
      default: false
      description: |
        Requests that the owner's own escrowed identity public key be
        included as an additional decrypt recipient, alongside whatever
        consumer recipients the client already resolved
        (specs/secret-objects/spec.md's "Opt-in owner-recipient inclusion
        at create time" requirement). Omitting this field (the default)
        means the owner is not added as a recipient - this service never
        makes ownership imply decrypt access on its own.

        This is a client-side sealing instruction, not something this
        service enforces or verifies: it never decrypts, and never adds a
        recipient to `value` itself - the caller has to add the owner's
        public key (`GET /auth/identity`) to its own recipient list before
        sealing. A create or update response echoes back exactly what that
        same request asked for; it isn't persisted, so a later fetch never
        carries this field.
    ObjectMetadata:
      type: object
      required: [slug, tags]
      properties:
        slug:
          $ref: "#/components/schemas/ObjectSlug"
        used_by:
          $ref: "#/components/schemas/UsedByList"
        tags:
          $ref: "#/components/schemas/Tags"
        description:
          $ref: "#/components/schemas/ObjectDescription"
        keep_readable_copy:
          $ref: "#/components/schemas/KeepReadableCopy"
        created_at:
          type: string
          format: date-time
          description: |
            When the object was created. Returned by `GET /objects`; left
            out of a create or update response.
        updated_at:
          type: string
          format: date-time
          description: |
            When the object's value was last replaced - equal to
            `created_at` for an object never updated. Returned by
            `GET /objects`; left out of a create or update response.
        created_by:
          allOf:
            - $ref: "#/components/schemas/Actor"
          description: |
            Who created the object, from the audit log's create entry. Left
            out when the audit log has none, as for an object that predates
            auditing.
        updated_by:
          allOf:
            - $ref: "#/components/schemas/Actor"
          description: |
            Who last updated the object, from the audit log. The creator
            for an object never updated. Left out when the audit log has
            no entry at all.
    Actor:
      type: object
      required: [type, id]
      description: |
        Who performed an audited write. `type` is `session` (the admin
        account, `id` is its actor id), `token` (a write token, `id` is its
        token id) or `consumer_token` (a consumer token, `id` is its id) -
        the same values `GET /audit-log` reports.
      properties:
        type:
          type: string
          examples: [session]
        id:
          type: string
          examples: [a1b2c3d4e5f6a7b8]
    Tag:
      type: string
      maxLength: 32
      pattern: "^[a-z0-9._/-]{1,32}$"
      examples: [prod]
    Tags:
      type: array
      maxItems: 10
      description: |
        Labels for grouping and filtering objects (`GET /objects?tag=`).
        Free-form, 1 to 32 characters from `a-z 0-9 . _ / -`, at most 10
        per object. Uppercase is converted to lowercase and duplicates are dropped.
        Metadata only - never part of the sealed value. A response always
        carries the array, empty when the object has none.
      items:
        $ref: "#/components/schemas/Tag"
      examples:
        - [prod, homelab]
    ObjectDescription:
      type: string
      maxLength: 1000
      description: |
        A free-text label set at creation, for a reader who only knows the
        slug. Fixed at creation - there is no way to change it later.
      examples: ["prod deploy webhook for homelab/vps-docker"]
    UsedByList:
      type: array
      maxItems: 100
      description: |
        The consumers (repos or hosts) recorded as depending on this
        object. Set at creation, and replaceable later via
        UpdateObjectRequest's own used_by field - a plain value update
        that omits it leaves the list as it was.
        A consumer named more than once is recorded once.
      items:
        type: string
        maxLength: 128
      examples:
        - [homelab/vps-docker]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: |
            What went wrong, in terms safe to show a caller. Nothing here
            handles authentication with a detail worth holding back.
          examples: [unknown object]
...